// src/core/interfaces/models/keyboard.interface.ts
export interface IKeyboard {
    piano: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    large: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    medium: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    small: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
}